data, yaml: move two functions into new #data-checks import
Due to import time shenanigans, these are provided bindFind and
getAllFindSpecs externally. Sorry!!
